+++ src/Text/Show/Text/Data/Monoid.hs
@@ -1,4 +1,7 @@
 {-# LANGUAGE CPP, TemplateHaskell #-}
+#if MIN_VERSION_base(4,8,0)
+{-# LANGUAGE FlexibleContexts #-}
+#endif
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-|
 Module:      Text.Show.Text.Data.Monoid
@@ -18,6 +21,9 @@ module Text.Show.Text.Data.Monoid (
     , showbLastPrec
     , showbProductPrec
     , showbSumPrec
+#if MIN_VERSION_base(4,8,0)
+    , showbAltPrec
+#endif
     ) where
 
 import Data.Monoid (All, Any, Dual, First, Last, Product, Sum)
@@ -29,6 +35,12 @@ import Text.Show.Text.Classes (Show(showbPrec), Show1(showbPrec1))
 import Text.Show.Text.Data.Bool ()
 import Text.Show.Text.Data.Maybe ()
 import Text.Show.Text.TH.Internal (deriveShowPragmas, defaultInlineShowbPrec)
+
+#if MIN_VERSION_base(4,8,0)
+import Data.Monoid (Alt)
+
+import Text.Show.Text.TH.Internal (mkShowbPrec)
+#endif
 
 #include "inline.h"
 
@@ -67,6 +79,13 @@ showbSumPrec :: Show a => Int -> Sum a -> Builder
 showbSumPrec = showbPrec
 {-# INLINE showbSumPrec #-}
 
+#if MIN_VERSION_base(4,8,0)
+-- | Convert an 'Alt' value to a 'Builder' with the given precedence.
+showbAltPrec :: Show (f a) => Int -> Alt f a -> Builder
+showbAltPrec = showbPrec
+{-# INLINE showbAltPrec #-}
+#endif
+
 $(deriveShowPragmas defaultInlineShowbPrec ''All)
 $(deriveShowPragmas defaultInlineShowbPrec ''Any)
 $(deriveShowPragmas defaultInlineShowbPrec ''Dual)
@@ -74,6 +93,11 @@ $(deriveShowPragmas defaultInlineShowbPrec ''First)
 $(deriveShowPragmas defaultInlineShowbPrec ''Last)
 $(deriveShowPragmas defaultInlineShowbPrec ''Product)
 $(deriveShowPragmas defaultInlineShowbPrec ''Sum)
+#if MIN_VERSION_base(4,8,0)
+instance Show (f a) => Show (Alt f a) where
+    showbPrec = $(mkShowbPrec ''Alt)
+    {-# INLINE showbPrec #-}
+#endif
 
 instance Show1 Dual where
     showbPrec1 = showbPrec