+++ src/Text/Show/Text/GHC.hs
@@ -10,6 +10,7 @@ Portability: GHC
 Imports 'Show' instances for @GHC@ modules.
 -}
 module Text.Show.Text.GHC () where 
+
 #if defined(mingw32_HOST_OS)
 import Text.Show.Text.GHC.Conc.Windows ()
 #endif
@@ -19,6 +20,10 @@ import Text.Show.Text.GHC.Event        ()
 # endif
 import Text.Show.Text.GHC.Fingerprint  ()
 import Text.Show.Text.GHC.Generics     ()
+#endif
+#if MIN_VERSION_base(4,8,0)
+import Text.Show.Text.GHC.RTS.Flags    ()
+import Text.Show.Text.GHC.StaticPtr    ()
 #endif
 #if MIN_VERSION_base(4,5,0)
 import Text.Show.Text.GHC.Stats        ()