+++ src/Text/Show/Text/Numeric.hs
@@ -1,0 +1,13 @@
+{-|
+Module:      Text.Show.Text.Numeric
+Copyright:   (C) 2014 Ryan Scott
+License:     BSD-style (see the file LICENSE)
+Maintainer:  Ryan Scott
+Stability:   Experimental
+Portability: GHC
+
+Imports 'Show' instances for @Numeric@ modules.
+-}
+module Text.Show.Text.Numeric () where
+
+import Text.Show.Text.Numeric.Natural ()